Move CollectionSidebar component to own dir
