+++ frontend/src/metabase/components/CollectionLanding.jsx
@@ -1,13 +1,24 @@
 /* eslint-disable react/prop-types */
 import React from "react";
 import { Box } from "grid-styled";
+import styled from "styled-components";
 
 import CollectionContent from "metabase/collections/containers/CollectionContent";
-import CollectionSidebar from "metabase/collections/containers/CollectionSidebar";
+import CollectionSidebar from "metabase/collections/containers/CollectionSidebar/CollectionSidebar";
 
 import * as Urls from "metabase/lib/urls";
 
 import { PageWrapper } from "metabase/collections/components/Layout";
+
+const sidebarWidth = "340px";
+
+const ContentBox = styled(Box)`
+  background-color: white;
+  height: 100%;
+  margin-left: ${sidebarWidth};
+  overflow-y: auto;
+  padding-bottom: 64px;
+`;
 
 const CollectionLanding = ({ params: { slug }, children }) => {
   const collectionId = Urls.extractCollectionId(slug);
@@ -15,19 +26,17 @@ const CollectionLanding = ({ params: { slug }, children }) => {
 
   return (
     <PageWrapper>
-      <CollectionSidebar isRoot={isRoot} collectionId={collectionId} />
+      <CollectionSidebar
+        isRoot={isRoot}
+        collectionId={collectionId}
+        width={sidebarWidth}
+      />
       {/* For now I'm wrapping this here so that we could potentially reuse CollectionContent without
         having the specific page margin and layout concerns, TBD whether that's a good idea or needed
         */}
-      <Box
-        bg="white"
-        className="border-left full-height"
-        style={{ overflowY: "auto" }}
-        ml={340}
-        pb={4}
-      >
+      <ContentBox className="border-left">
         <CollectionContent isRoot={isRoot} collectionId={collectionId} />
-      </Box>
+      </ContentBox>
       {
         // Need to have this here so the child modals will show up
         children
